create init

+++ database/migrations/2017_12_28_105738_create_ingredients_table.php
@@ -17,7 +17,7 @@ class CreateIngredientsTable extends Migration
             $table->increments('id');
             $table->string("name");
             $table->char('unit');
-            $table->integer('heat');
+            $table->float('heat');
             $table->timestamps();
         });
     }
